backend/requirements.txt
# Core requirements
psycopg2-binary==2.9.1
Flask==2.0.1
flask-login==0.5.0
flask-mail==0.9.1
flask-mysqldb-1.0.1
flask-cors==3.0.10
